add file loader, fix loaders and static serve
