+++ src/components/LoginForm/LoginForm.js
@@ -1,0 +1,3 @@
+/**
+ * Created by levithomason on 2/7/16.
+ */
