+++ src/views/ApplyView/LoginView.scss
@@ -1,0 +1,10 @@
+html,
+body {
+  background: lightgray;
+}
+
+.LoginView {
+  &--full-height {
+    height: 100vh;
+  }
+}
