--- src/views/HomeView/HomeView.scss
@@ -1,8 +1,0 @@
-.loginColumn {
-  position: absolute;
-  top: 20%;
-  right: 0;
-  left: 0;
-  max-width: 340px;
-  margin: auto;
-}
